+++ tests/common.js
@@ -2,11 +2,35 @@ import {
   deviceInfo155 as deviceInfo,
   mockListAppsResult,
 } from "@ledgerhq/live-common/lib/apps/mock";
+import { Application } from "spectron";
+import _ from "lodash";
 import { toMatchImageSnapshot } from "jest-image-snapshot";
-import { applicationProxy, removeUserData, getMockDeviceEvent } from "./applicationProxy";
 import ModalPage from "./po/modal.page";
 import fs from "fs";
+import rimraf from "rimraf";
 import path from "path";
+
+Application.prototype.startChromeDriver = function() {
+  this.chromeDriver = {
+    start: () => {
+      return Promise.resolve();
+    },
+    stop: () => {
+      return Promise.resolve();
+    },
+    clearLogs: () => {
+      return [];
+    },
+    getLogs: () => {},
+  };
+  return this.chromeDriver.start();
+};
+
+const getMockDeviceEvent = app => async (...events) => {
+  return await app.client.execute(e => {
+    window.mock.events.mockDeviceEvent(...e);
+  }, events);
+};
 
 let app;
 let modalPage;
@@ -17,8 +41,63 @@ jest.setTimeout(600000);
 
 // eslint-disable-next-line jest/no-export
 export default function initialize(name, { userData, env = {}, disableStartSnap = false }) {
+  const userDataPathKey = Math.random()
+    .toString(36)
+    .substring(2, 5);
+  const userDataPath = path.join(__dirname, "tmp", userDataPathKey);
+
+  const removeUserData = dump => {
+    if (fs.existsSync(`${userDataPath}`)) {
+      if (dump) {
+        fs.copyFileSync(`${userDataPath}/app.json`, path.join(__dirname, "dump.json"));
+      }
+      rimraf.sync(userDataPath);
+    }
+  };
+
   beforeAll(async () => {
-    app = await applicationProxy(userData, env);
+    fs.mkdirSync(userDataPath, { recursive: true });
+
+    env = Object.assign(
+      {
+        MOCK: true,
+        DISABLE_MOCK_POINTER_EVENTS: true,
+        HIDE_DEBUG_MOCK: true,
+        DISABLE_DEV_TOOLS: true,
+        SPECTRON_RUN: true,
+      },
+      env,
+    );
+
+    if (userData) {
+      const jsonFile = path.resolve("tests/setups/", `${userData}.json`);
+      fs.copyFileSync(jsonFile, `${userDataPath}/app.json`);
+    }
+
+    app = new Application({
+      path: require("electron"), // just to make spectron happy since we override everything below
+      waitTimeout: 15000,
+      webdriverOptions: {
+        capabilities: {
+          "goog:chromeOptions": {
+            binary: "/app/node_modules/spectron/lib/launcher.js",
+            args: [
+              "spectron-path=/app/node_modules/electron/dist/electron",
+              "spectron-arg0=/app/.webpack/main.bundle.js",
+              "--disable-extensions",
+              "--disable-dev-shm-usage",
+              "--disable-gpu",
+              "--no-sandbox",
+              "--lang=en",
+              `--user-data-dir=/app/tests/tmp/${userDataPathKey}`,
+            ].concat(_.map(env, (value, key) => `spectron-env-${key}=${value.toString()}`)),
+            debuggerAddress: undefined,
+            windowTypes: ["app", "webview"],
+          },
+        },
+      },
+    });
+
     modalPage = new ModalPage(app);
     mockDeviceEvent = getMockDeviceEvent(app);
